Move devices to a new package
