+++ backend/src/main/java/pt/ua/deti/ies/SmartHomes/backend/Devices/DeviceType.java
@@ -1,4 +1,4 @@
-package pt.ua.deti.ies.SmartHomes.backend.Database;
+package pt.ua.deti.ies.SmartHomes.backend.Devices;
 
 enum DeviceType {
     SOLARPV, WINDTURBINE, PLUG, GRIDMETER, AC, LIGHT, TV, SMARTASSISTANT, CARCARGER, VACUUM, DESUMIDIFIER, OVEN
